Build: Add font generation to the grunt process
This automates the build process. To generate a font, now make sure you've run `npm update`, then run `npm run build`.

The `codepoints.json` (props @field2) should ensure that existing icons are mapped to the same unicode locations as the old font was. But we need to verify this thoroughly.

Now that the font generation is automated, we should consider retiring/deleting these folders:

- source
- js
- fonts
- css

At a glance it looks like all of those are obsolete now. Is that correct?