Add icon: Full-width
This PR seeks to add a new "full width" icon. The Gutenberg editor needs this for a new layout control option for things like images and embeds. When an image or embed is set to this layout option, it breaks out of the main text column and fills 100% of the width of the browser viewport.