Subtly redesin the Trash icon
In this PR I'd like to propose some small tweaks to the trash icon. There are two purposes to this:

1. Allow it to scale better
2. Reduce the overall silhuette

Both of these come as a result of using the trash icon next to blocks in the Gutenberg editor project. For every block you click, you'll see a trash can you can click to remove that block.

By reducing the complexity of the shape, we make that visual load lighter.

An additional trick, is scaling the icon to 75%. Scaling outside of the pixel grid is usually a no-no, but there are increments where pixels look their best when scaled. This usually means 2x and 3x, but .75x and .5x can work too. For Gridicons, we are pondering embracing .75x too. That set is at a 24px base grid, so .75x would give us 18px. Many icons, especially solids, scale well to this size without further work, and we are looking into a way to load individually hinted versions for that size, for the icons that need it.

For Dashicons this means we can possibly make some scaling exceptions, and show a few icons, usually solids, at 15px, and it'll still look pretty good. This would further reduce the visual load of having the trash button near every block in Gutenberg.