Try adding Tide icon.
See #248.

This builds and doesn't break the build process. Not sure why the old icon broke it, except for that it was a 14x14px SVG that wasn't a compound shape.

Opening the initial SVG in Illustrator, copying the artwork to a new 20x20px document, making it a compound shape with the pathfinder, made it work.