Add helper SCSS file to React component
This adds a very simple SCSS file to the React component, with a few rules for any SVG icon that uses the `dashicon` class.

The fill rule makes sure the icon inherits the `color` of the parent element. This means for example that if you put a dashicon inside a link, the dashicon will have the same blue color as the link, even the purple color when visited.

The outline rule ensures that the icon is uses as just that, a decorative element, that inherits the focus style of the parent.