Add Cover Image icon
Right now in Gutenberg, the icon for the Cover Image block is the same as that for the Image block. This PR adds a new icon for the Cover Image block, fixing https://github.com/WordPress/gutenberg/issues/5516#issuecomment-372048002. The idea is based on the icon for full-wide, which shows text below the image in a more widescreen format, indicating that it's a "bolder" image.

Thoughts?

If you like it I need to run grunt first.