Fix the eyes.
This PR outlines the two eye-icons, and rebuilds them. This fixes #392.

Of course the newly built TTF font also needs to be added to WordPress core, but as far as the Dashicons repo is concerned, it fixes that issue.

It also updates a number of dependencies that were out of date. See packagge*json files.